redesigned citation dialog
- Implemented redesigned citation dialog with library
and list modes one can switch between.
- one can add any locator (not just pages) by typing it's full or short name in an input and pressing Enter
(e.g. line 10, or l. 10, or chap. "test chapter")
- Added a new preference to select if the citation dialog
should always open in list mode, library mode or in the last
mode that was used
- bubbles whose items are selected in library or list mode
are highlighted
- arrow up/down from a bubble will focus the bubble
above/below it for easier navigation across bubble-input

In Library Mode:
- searching will filter the items table
- items table has a new + button column to add items from the
citaiton and rows of selected items are highlighted
- open, cited and selected items appear in a section
between the items table and bubbleInput
- only top-level items are shown when citing
items and only notes/notes' parents - when adding a note
- selected items are initially grouped together and clicking them will add all of them.
But they can be expanded by clicking on "Expand all"
- when an item is added by clicking on the + icon, bubble-input
may increase in height and push itemTree lower. To try to preserve
relative positioning of the mouse, itemTree will scroll to be over
the row that was just clicked

In List mode:
- arrow up/down from the input will change the selected item
with the focus remaining in the input. Screen readers will
announce the items because we update aria-activedescendant.
- selected items are a collapsible list section

Implementation notes:
- citationDialog.js is the main file. It relies on a number of
helper files in citaitonDialog/* directory to keep the main
file less cluttered. popupHandler.js contains the logic of
opening/closing the item details popup to add locator/prefix/suffix/etc.
keyboardHandler.js is responsible for overall keyboard navigation
throughout the dialog. searchHandler.js contains the logic
for running the search based on user's query. Finally, Helpers.js
has general helper functions that don't handle any actual logic.
- bubbleInput.js is a customElement responsible for bubbles interface.
bubbleInput.refresh takes a list of items and handles adding/removing/reordering
of bubbles as needed
- refactored bubbleInput to always have two inputs on each side of
a bubble, as opposed to having inputs inserted dynamically. It
allows keyboardHandler.js to handle navigation with arrows within
bubbleInput.
- when user interacts with bubbleInput, it emits custom events that are
handled by IOManager singleton in citationDialog.js, which
may upate the items information and pass them back to bubbleInput.refresh
to have the list of bubbles updates
- CitationDataManager singleton is responsible for storing
items added into the citation in CitationDataManager.items
in an objct with both Zotero.Item and the actual citation item.
It is easier to pass both items to other components
and helpers, as opposed to sharing functions to convert
items back and forth.
- added getExtraField prop to itemTree to get data on if
an item is in a citation or not
- added initialFolder and onActivate prop to collectionTree to set which collection
should be reopened when dialog opens and to be able to set a custom
onActivate handler.