a11y edits
- revised instructions for the screen readers
- added instructions to collectionTree and itemTree
on what they are for, and which keypresses are available
- all inputs have a default placeholder. For some reason,
it is necessary for voiceover to properly place virtual
cursor on them, as opposed to bubble-input wrapper.
- removed earlier combobox setup, where selected items
were set as active descendents of the combobox input.
If you tab into the input, it does not reliably read out
the aria description of the inputs themselves, which
almost soudns like you tabbed onto an item, not an input.
Also, it does not makes sense in library mode. So probably
best to just remove it, and now we announce similar to QF
that one should Tab to navigate matching items after
typing a query.
- aria props for collapsible sections buttons and
item details popup