optimize context menus with collections
- more efficiently check if a collection can be moved
to another collection without calling getDescendents
for every single candidate
- when fetching descendents of a collection, only fetch
those of the relevant type

- It turns out that adding many top-level menuitems
with icons into a menupopup takes a while to
process and may freeze the UI on macOS. This issue only emerges
when icons are present. To avoid it, the icons are hidden
for all but the first few menuitems and then are dynamically
added when the browser is idle.

Fixes: #4963