render more icons + avoid changing menu width
- increase number of initially rendered icons to 75
- do not optimize icons rendering if there are < 250
menuitems. Just let all icons render normally as the
delay is not too big at that point.
- do not remove the icon from the longest menuitem
to make it more likely that the menu will occupy the
correct width from the start
- cache and re-use the "New collection" menuitem
at the top. Since doc.l10n is asynchronous, the menuitem's
label is set after initial icons are rendered, which
means the label is empty for a moment and looks off. It also
affects the width of the menu if all collections are
short, so when "New collection" is set, it may increase
menu's width. Reusing the created row avoids these issues.
- fix unwanted repeated call to showMenuIconsOnIdle from
"Move to" collection menu