refactor locator extraction
Redo logic of extracting the locator to allow for better
detection of locators and to not rely on longer regex that
are harder to troubleshoot.

These are the new rules for extracting locators:
- locator value can be either a set of numbers potentially
with some punctuation in between OR text surrounded with
single or double quotes
- locator labels can be typed in any of 3 formats:
full (chapter), short (chap.), or short with no
punctuation (chap)
- locator labels go before the locator values and there
may or may not be a space in between. For example,
is the same as chap.10, chap. 10.