citation dlg: earlier init for accept/cancel func
Added IOManager.preInit that is ran as soon as the dialog
is loaded. Move accept/cancel initialization there.

IOManager.init runs at the very end of onLoad because
it relies on most data and layouts being loaded,
It means if there is some delay or error during loading,
the dialog cannot be closed. To avoid it, setup
cancel/accept buttons separately in the beginning.