sciencedirect: potential fix to infinite captcha
Before opening the PDF, open sciencedirect homepage
where a number of cloudflare cookies (including cf_clearance
indicating that this is a real user) are set. Wait
for the cookies to be added to the sandbox, and then
proceed to open the PDF. In that case, captcha should
not appear at all.