collection search: keep children of matches
- keep children of matching collections in the tree
so one can access them after looking up the parent.
- top-most collection matching the filter is collapsed,
so the children do not clutter the tree.
- comment out arrow up/down navigation skipping over
non-matching rows. Otherwise, there would be no way
to access non-matching children via keyboard.
- rework how the current collectionTreeRow is preserved
on the itemTree. Do not add a hidden collectionRow
to the bottom of the tree to avoid edge cases
related to having a hidden row with a real row potentially
being present in the tree as a child of collapsed parent.
Instead, clear the selection during filtering so
it does not trigger itemTree selection changes and
does not affect selection when rows are collapsed.