expand container rows on dragover after delay
- when something is dragged over a container, do not
expand it immediately to avoid opening all containers
in the way of a dragged item. Instead, wait a second
and expand the container if a second later, it is
still being dragged over.
- when items are dragged out of the collectiomTree,
opened containers will be closed after a second too.
It is prevented in case if one drags an item from
the collectionTree and then drags it back into it
immediately.
- added a flashing effect right before a container will
open