itemTree performance optimization in search mode
When an item is modified in search mode, do not rerun
the entire query and refresh the entire search list.
Instead, run the query on the modified items and manually
add/remove/expand/collapse item rows depending on if
updated items and their parents/children should still
be present in search results.