updated itemTree manual row processing in notify
- collapse all rows before removing them. It makes
sure that annotation rows will be cleaned up properly
when an annotation row is moved into another parent.
- refactor conditionals to handle changes of parent
regardless of if the item is a container (attachment
with annotations) or not (a note).
- added tests for itemTree's handling of changing the
parent of attachments and notes, as well as ensuring
there are no ghost annotation rows in Unfiled Items.

Fixes: zotero#5246