enable adding annotation rows into note editor
- set zotero/annotation data when annotation rows are
dragged into the note editor, so that it can add a
citation, the same way as when dragging from the reader
- when copying annotations to the clipboard as a note,
set zotero/annotation data so that note-editor can handle it.
zotero/annotation does not appear by itself in event.clipboardData on
paste, so it will be passed into note-editor iframe in
window.clipboardData object, similar to how drop events
are handled
- tweak Zotero.Annotations.toJSONSync to optionally
add fields used to serialize annotations in
Zotero.EditorInstanceUtilities.serializeAnnotations.
Set annotation.id to be item.key, which will add
the &annotation= param into the link to the zotero://open-pdf/
link after the citation.

Fixes: #5235