citationDlg: more optimization of initial loading
https://forums.zotero.org/discussion/124211/zotero-update-bug-delay-in-citation-insertion-search

It can take a while for all the promises that io needs to sort to resolve.
It means that if we await for sorting while building citation
on initial load, it can cause a visible delay.

When an item is added, only run sorting if io is ready for it.
If io is not ready to sort during initial load, some
bubbles may be added out of order. They will be sorted
as soon as io is ready.

Also, add a few Zotero.debug lines.