persist open state of attachment rows
- record in a persistent cache which itemTree rows
were toggled open. When a row is being expanded, check
if its children are meant to be open, and if so - expand
them as well
- that way, if one expands an attachment, collapses
the parent item, and then expands the parent item, the
annotations will be visible
- if an attachment is expanded and its annotations
are all removed in the reader, it will technically count
as a closed container. But if some new annotations are added later,
the attachment will reopen. This also applies to
attachments with just one annotation that is converted
between highlight and underline, since it essentially
deletes the old annotation and creates a new one
- when a child item (e.g. note) is modified, the parent
is collapsed and re-opened during sorting. If another
child item is an expanded attachment, this will ensure
that it remains open
- for consistency, if an attachment is expanded and
is moved to/from top-level items, it will remain open

https://forums.zotero.org/discussion/123933/persistent-expand-collapse-state-for-annotations-in-the-items-list
https://forums.zotero.org/discussion/123934/convert-between-highlight-and-underline-hides-the-annotation-from-the-items-list

Fixes: #5252