server connector endpoint to cancel a session
- added /connector/cancelSession endpoint to connector server that
the connector can call to cancel or undo the saving process.
- if there is not session, it will be created, so we can
record that the session with that ID is canceled for
further endpoint calls.
- when the session is cancelled, subsequent calls
to /saveAttachment, /saveSnapshot, /saveSingleFile,
/saveItems, /updateSession are rejected with 409 error.
- after the session is cancelled, wait a second,
and try to remove any existing session items again.
This accounts for edge cases like if an items is saved at the
same time as the session is cancelled.

Additionally, fixed session not getting removed or garbage-collected