tag autocomplete matching non-left bound substring
Tag autocomplete will now match tag substrings in the middle,
not just the start of the tag.

Refactoring to optimize performance for libraries
with many tags:
- From each search, save results into an in-memory cache.
During subsequent searches, check the cache to see if
it has any matches
- Ensure we never run multiple queries at the
same time. If a query is already running, wait for it to
finish before running the next one. If there are multiple
pending queries that arrive while one is running,
discard all but the last one
- Fixed stopSearch not working correctly. Now, if
stopSearch is called, the current query processing will be
properly cancelled in the onRow handler.

Also, drop resultsCallback to keep function flow clearer, since it
is not used anywhere at the moment.