annotation color/author filters in tag selector
- added new component annotationFiltersSelector.jsx to
display annotation colors followed by a list of annotation
authors
- a splitter between the annotations filters and
the actual tag selector can be used to adjust the heights
of these two subsections
- search box of the tag selector will also filter
the annotation authors
- added new search conditions to accomodate searching
by annotation color and author
- the logic of which tag colors/authors are visible
is somewhat different from tags, and a bit more convoluted.
Filtering by many annotation colors or authors is done
with OR operator, not AND as with tags. It means that
if one annotation author is selected, other annotation
authors should not be removed from the list of options, so
that one can select more than one author.
So the current search logic can be described as below:
(tag1 AND tag2) AND (color1 OR color2) AND (author1 OR author2)
- to be able to determine which colors and authors
can be selected, we need to "ignore" any existing colors
and authors conditions. (Since once a single author is
selected, no other author will be present in the scope).
To achieve this, collectionTreeRow.getSearchObject supports
{ skipAnnotationAuthors, skipAnnotationColors } options that
will construct the search object without these conditions.