support adding annotation items as a mock note
- update editor instance to be able to create a note
from annotations without saving the note into the
database, which gives us the temporary mock note that can
inserted from the citation dialog
- allow ink annotations to be inserted into notes as
well as images for consistency
- when the image or ink annotations are being inserted,
they cannot be added into the note via attachmentKey
(since it would require to save entries to database).
Instead, they are embedded as a src="data:image/..." string
- display selected annotations in a deck separate from
selected items
- adding an annotation will bubbleize it
- clicking on a selected/open/cited top-level item will
select that item in itemTree
- list mode is hidden for now in "add annotations" mode
- side pane to the right of itemTree will display
a preview of selected annotations or all annotations
of a selected item. Added + icon to annotation-rows
to add annotations from the side pane.
- added annotations are groupped by their attachment
and sorted by annotationSortIndex. Reordering of bubbles
is disabled

Temporarily, one the citaiton dialog in "Add citation" mode
when "Add note" is clicked.