remove duplicate cslEngine initialization
Instead, when the dialog opens, record if there are
any omitted items, clear them and generate the bibliography
as of nothing is omitted. That way, we can display omitted
items at their proper position.

When the dialog is accepted, omittedItemIDs may have been
modified, so they will be removed from the bibEditInterface
again.