citation dlg: updates to locators shortcuts
- after a new bubble is added to the citation, it is recorded
as a just-added bubble. The next locator typed without any additional
search query and accepted via Enter will go to that item,
instead of going to the bubble right before where the locator was typed.
- just-added bubble is visually marked, so it is more
apparent why a locator would go into that bubble.
- the record of just-added bubbles is cleared on focusout
or keypress of an arrow key. That way, it's discarded if the user
is almost certainly not intending to immediately type a locator.
- added a special case to recognize a numeric value as a page locator
if it is typed right after a bubble is added
- added a special case to recognize ":<number>" as a locator
in the same circumstances that "page <number>" is currently recognized

Also, minor refactoring of cleanSearchQuery to make it less
likely that relevant punctuation (e.g. colon for locators) is removed.