citation dlg: updates to locators shortcuts
- after a new bubble is added to the citation, it is recorded
as a just-added bubble. The next locator typed without any additional
search query and accepted via Enter will go to that item,
instead of going to the bubble right before where the locator was typed
- the record of just-added bubbles is cleared on focusout
or keypress of an arrow key. That way, it's discarded if the user
is almost certainly not intending to immediately type a locator.
- added a special case to recognize ":<number>" as a locator
in the same circumstances that "page <number>" is currently recognized
- added a special case to recognize a numeric value as a page locator
if it is typed right after a bubble is added. That special locator
will be added to the just-added bubble as one is typing without
pressing Enter. This is to preserve previous behavior from
quickFormat dialog.
- if a just-added bubble is recorded, cmd-z will clear
whatever numeric locator may have been typed and place
it back into the input, in case one meant to type an
actual search query
- do not use year extraction (SearchHandler._cleanYear)
when parsing input. It strips the first number from
a range of numbers and conflicts with the new locator
logic.
- ensure a bubble with a very long locator does not overflow